third_party/blink/web_tests/http/tests/history/cross-origin-replace-history-object-child-expected.txt

ALERT: PASS: Access to window.frames[0].history threw an exception.
CONSOLE MESSAGE: line 17: About to shadow child window's history object: [object History]
CONSOLE MESSAGE: line 19: Child window's history object should not be shadowed: [object History]
ALERT: PASS: Access to window.frames[0].history threw an exception.
